[UI] Phase 5.a: Split ConnectionTable into focused siblings (#18660)
components/connections/ConnectionTable.tsx was 1428 significant lines,
over the 1000-line hard ceiling tracked by `npm run audit:size` (§7.5
of the UI restructure plan). Splits it into focused sibling files;
entry point drops below 600 lines and `over_hard` drops by 1.

New sibling files in ui/components/connections/:
  ConnectionTable.types.ts        shared types (ConnectionRow, props,
                                   filters, expansion flags)
  ConnectionTable.constants.ts    ACTION_TYPES, transition map,
                                   getErrorMessage / getStatusTransition,
                                   docs URLs
  ConnectionTable.columns.tsx     useConnectionColumns() — the large
                                   columns useMemo with all body/head
                                   renderers
  ConnectionTable.options.tsx     useConnectionTableOptions() — the
                                   ResponsiveDataTable options useMemo
                                   (toolbar select, sort handler,
                                   expandable rows)
  ConnectionTable.hooks.ts        useConnectionActions() — RTK Query
                                   mutation handlers (env assign/save,
                                   status update)
  ConnectionTableToolbar.tsx      search bar + universal filter +
                                   column visibility toolbar row
  ConnectionActionMenu.tsx        the kubernetes-only action Popovers
                                   (Flush MeshSync + deployment mode)

Entry-point default export, prop signature and rendered DOM are
unchanged. ESLint and tsc clean; legacyMaxLineOffenders no longer
lists ConnectionTable.tsx. ConnectionTable.columns.tsx and
ConnectionActionMenu.tsx are added to legacyRestrictedImportOffenders
because they retain the existing `@mui/icons-material/*` imports from
the original (per PR scope — no MUI migration in this slice).

Refs #18660
Refs #18654
