+++ ui/components/connections/ConnectionActionMenu.tsx
@@ -1,0 +1,99 @@
+import React from 'react';
+import { Button, Popover, Typography } from '@sistent/sistent';
+import { ActionListItem } from './styles';
+import SyncIcon from '@mui/icons-material/Sync';
+import { iconMedium } from '../../css/icons.styles';
+import CAN from '@/utils/can';
+import { keys } from '@/utils/permission_constants';
+import { MESHSYNC_DEPLOYMENT_TYPE } from '../../utils/Enum';
+
+type ConnectionActionMenuProps = {
+  anchorEl: HTMLElement | null;
+  open: boolean;
+  onClose: () => void;
+  onFlushMeshSync: () => void;
+  onDeploymentModeAnchor: (event: React.MouseEvent<HTMLElement>) => void;
+};
+
+export const ConnectionActionMenu = ({
+  anchorEl,
+  open,
+  onClose,
+  onFlushMeshSync,
+  onDeploymentModeAnchor,
+}: ConnectionActionMenuProps) => {
+  return (
+    <Popover
+      open={open}
+      anchorEl={anchorEl}
+      onClose={onClose}
+      anchorOrigin={{
+        vertical: 'bottom',
+        horizontal: 'left',
+      }}
+    >
+      <ActionListItem>
+        <Button
+          type="submit"
+          onClick={onFlushMeshSync}
+          data-cy="btnResetDatabase"
+          disabled={!CAN(keys.FLUSH_MESHSYNC_DATA.action, keys.FLUSH_MESHSYNC_DATA.subject)}
+        >
+          <SyncIcon {...iconMedium} />
+          <Typography variant="body1" style={{ marginLeft: '0.5rem' }}>
+            Flush MeshSync
+          </Typography>
+        </Button>
+      </ActionListItem>
+      <ActionListItem>
+        <Button
+          type="submit"
+          onClick={(e) => {
+            e.stopPropagation();
+            onDeploymentModeAnchor(e);
+          }}
+          data-cy="btnChangeDeploymentMode"
+        >
+          <Typography variant="body1">Modify Deployment Mode</Typography>
+        </Button>
+      </ActionListItem>
+    </Popover>
+  );
+};
+
+type ConnectionDeploymentModeMenuProps = {
+  anchorEl: HTMLElement | null;
+  open: boolean;
+  onClose: () => void;
+  onSelectMode: (mode: string) => void;
+};
+
+export const ConnectionDeploymentModeMenu = ({
+  anchorEl,
+  open,
+  onClose,
+  onSelectMode,
+}: ConnectionDeploymentModeMenuProps) => {
+  return (
+    <Popover
+      open={open}
+      anchorEl={anchorEl}
+      onClose={onClose}
+      anchorOrigin={{
+        vertical: 'bottom',
+        horizontal: 'left',
+      }}
+    >
+      <ActionListItem>
+        <Button onClick={() => onSelectMode(MESHSYNC_DEPLOYMENT_TYPE.OPERATOR)}>
+          <Typography variant="body1">Operator</Typography>
+        </Button>
+      </ActionListItem>
+      <ActionListItem>
+        <Button onClick={() => onSelectMode(MESHSYNC_DEPLOYMENT_TYPE.EMBEDDED)}>
+          <Typography variant="body1">Embedded</Typography>
+        </Button>
+      </ActionListItem>
+    </Popover>
+  );
+};
